Added queueCount property.
Added queueCount property. If a message is received with this property, node will send message with number of messages left in the queue in queueCount property. This message won't store in queue.
Now every message released by node, will have queueCount property with number of messages left in the queue at the moment.